Fix GameDrivesRepository ID type mismatch (Integer vs Long)

+++ backend/src/main/java/com/nfldb/game_drives/GameDrivesRepository.java
@@ -3,6 +3,6 @@ package com.nfldb.game_drives;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface GameDrivesRepository extends JpaRepository<GameDrives, Long> {
+public interface GameDrivesRepository extends JpaRepository<GameDrives, Integer> {
     List<GameDrives> findByGameId(String gameId);
 }
